refactor: Remove unused components and update component names for consistency
