--- app/(tabs)/home/filters.tsx
@@ -1,12 +1,0 @@
-import { View, Text } from "react-native";
-import React from "react";
-
-const Filters = () => {
-  return (
-    <View>
-      <Text>Filters</Text>
-    </View>
-  );
-};
-
-export default Filters;
