+++ app/_layout.tsx
@@ -77,8 +77,9 @@ const StackLayout = () => {
       <Stack.Screen name="index" options={{ headerShown: false }} />
       <Stack.Screen name="(auth)" options={{ headerShown: false }} />
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack.Screen name="(pages)" options={{ headerShown: false }} />
       {/* <Stack.Screen name="/search/[query]" options={{ headerShown: false }} /> */}
-      <Stack.Screen name="+not-found" />
+      <Stack.Screen name="+not-found" options={{ headerBackTitle: "Back" }} />
     </Stack>
   );
 };
